requirements.txt
openai
python-dotenv
azure-identity
requests
flask